refactor(executor_service): 拆分上帝模块为 9 个职责单一子模块 (issue #674)
将原 backend/src/executor_service.rs（2962 行、65 个函数、22 个超 30 行）
按 SRP 拆为 9 个子模块：

- mod.rs (140 行)           公共 facade + run_todo_execution 顶层 orchestrator
- worktree.rs (197 行)      WorktreeContext + Git Worktree 集成 (issue #643)
- log_capture.rs (438 行)   stdout/stderr reader + LogFlusher pipeline + stats
- pre_spawn.rs (538 行)     失败短路 reject_xxx + 并发上限 + executor 选择
- completion.rs (527 行)    终态分支（正常/取消/超时）+ completion hook + 事件
- spawn_lifecycle.rs (680 行) spawn 闭包生命周期（dispatch_outcome 三分支）
- stages.rs (297 行)        三阶段 stage 函数 + dispatch_spawned_executor_task
- types.rs (138 行)         跨模块 stage 产物聚合类型
- auto_review.rs (387 行)   同步自动评审（独立 runtime + std::thread）

关键抽取：
- run_auto_review_inner 156 行 → 7 个 ≤ 25 行 helper（load / validate / compose / execute / poll）
- spawn_stdout_reader 113 行 → 3 个 helper（session_id / progress / stats）
- dispatch_outcome 86 行 → 3 个 dispatch_* helper + 轻量 match wrapper
- select_executor_and_build_command 64 行 → extract_todo_executor_fields /
  resolve_executor_instance / build_executor_command_args / persist_executor_choice
- create_run_execution_record 61 行 → build_command_string + create_record_or_reject
- finalize_normal_completion 66 行 → maybe_run_auto_review + emit_completion_events

公共 API 不变：
- RunTodoExecutionRequest / ExecutionResult / run_todo_execution /
  run_todo_execution_with_params / run_auto_review 签名与字段完全一致
- 所有 5 个外部调用方（scheduler / hooks / webhook / message_debounce / execution handler）无需修改

测试：
- 单元测试 649 passed（含 20 个 executor_service 测试）
- 编译期签名断言（RunTodoExecutionRequest.hook_service / PreparedExecution /
  SpawnInputs / SpawnRuntime.effective_workspace）保留防回归功能

Refs: #674
